Added tests for all of the functions and changed the classes name(removed 'ForStrings' part)
